requirements.txt
python-dotenv==1.0.1
pydantic==2.8.2
pandas==2.2.2
notion-client==2.2.1
openai==1.45.0 

langchain==0.2.15
langchain-openai==0.1.20
langchain-community==0.2.3
langgraph==0.1.19
